requirements.txt
# Core ML and RL frameworks
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
peft>=0.4.0
trl>=0.7.0

# RL and optimization
gymnasium>=0.29.0
stable-baselines3>=2.0.0
optuna>=3.2.0

# Data processing and utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Code execution and verification
sympy>=1.12.0
astunparse>=1.6.3
black>=23.0.0
pylint>=2.17.0
ast-comments>=1.0.0

# Jupyter and development
jupyter>=1.0.0
ipywidgets>=8.0.0
tqdm>=4.65.0
rich>=13.0.0

# Logging and monitoring
wandb>=0.15.0
tensorboard>=2.13.0
mlflow>=2.4.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
pydantic>=2.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.4.0
types-requests>=2.31.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0 